Split runner process helpers into submodules
